Fix KeepAlive refresh
The state of every device was getting reset on every network refresh,
which is problematic when we want to display devices that weren't in the
refresh, but haven't passed their keepalive timeout.

This change determines if a device has gone past the timeout before
saving it to the devices to be displayed on the screen.